requirements.txt
matplotlib==3.10.3
numpy==2.3.2
pytorch==2.5.1
torchvision==0.20.1